requirements.txt
aiogram==2.25.1
openai==0.27.4
python-dotenv==1.0.0